Post Busselton updates
Update map2model c++ code to display mineral despoit tectonostrat relationships
Include multiscale code m2l_subsample.py of Ranee Joshi and associated notebook and test data
Tidy LoopStructural all in one notebook to show more incremental data products inline
Include test code in m2l_topology to use regional stratighic table to provide hints for formation ordering in save_geol_wkt with notebook 107 to generate required inputs
localise impoorts of gempy and LoopStructural withing relevant functions so libraries only needed if that fucntion is called
Remove redundant basal contact displays from m2l_interpolation
Add nice point plotting over geology to m2l_utils

+++ m2m_source/src/converter_types.h
@@ -26,7 +26,7 @@ public:
   // Unique id.
   int id;
   // Min/max age ma (for polygons).
-  int min_age, max_age;
+  double min_age, max_age;
   // Unit name (for polygons).
   std::string name;
   // Group name (for polygons).
